Change parseFloat to parseInt

+++ scripts.js
@@ -8,17 +8,17 @@ let output;
 let pressed;
 
 let add = (inputs1, inputs2) => {
-	return parseInt(inputs1) + parseInt(inputs2);
+	return parseFloat(inputs1) + parseFloat(inputs2);
 };
 let substract = (inputs1, inputs2) => {
-	return parseInt(inputs1) - parseInt(inputs2);
+	return parseFloat(inputs1) - parseFloat(inputs2);
 };
 let multiply = (inputs1, inputs2) => {
-	return parseInt(inputs1) * parseInt(inputs2);
+	return parseFloat(inputs1) * parseFloat(inputs2);
 };
 let divide = (inputs1, inputs2) => {
-	if (parseInt(inputs2) !== 0) {
-		return parseInt(inputs1) / parseInt(inputs2);
+	if (parseFloat(inputs2) !== 0) {
+		return parseFloat(inputs1) / parseFloat(inputs2);
 	}
 };
 
@@ -26,7 +26,7 @@ let getInput = () => {
 	let input = "44 + 55 / 66";
 	/* Put the inputs and operators in arrays */
 	inputs = input.match(/\d+/g);
-	inputs = inputs.map(x => parseInt(x));
+	inputs = inputs.map(x => parseFloat(x));
 	operators = input.match(/ \D/g);
 	/* Remove the whitespaces */
 	operators = operators.map(x => x.replace(/\s/, ""));
